+++ src/library/DIPS.Mobile.UI/Components/Buttons/iOS/ButtonHandler.cs
@@ -1,20 +1,28 @@
-using System.Runtime.CompilerServices;
 using CoreGraphics;
 using DIPS.Mobile.UI.Platforms.iOS;
-using Microsoft.Maui.Handlers;
 using Microsoft.Maui.Platform;
 using UIKit;
-using IImage = Microsoft.Maui.IImage;
 
 namespace DIPS.Mobile.UI.Components.Buttons;
 
 public partial class ButtonHandler : Microsoft.Maui.Handlers.ButtonHandler
 {
+    private int LargeButtonHeight { get; } = Sizes.GetSize(SizeName.size_12);
+    
     protected override UIButton CreatePlatformView()
     {
         var button = new UIButtonWithExtraTappableArea();
         return button;
     }
+
+    protected override void ConnectHandler(UIButton platformView)
+    {
+        base.ConnectHandler(platformView);
+
+        var percentOfHeight = (float)VirtualView.Height / LargeButtonHeight;
+        platformView.ImageView.Transform = CGAffineTransform.MakeScale(percentOfHeight, percentOfHeight);
+    }
+
     private partial void AppendPropertyMapper()
     {
     }
@@ -26,7 +34,9 @@ public partial class ButtonHandler : Microsoft.Maui.Handlers.ButtonHandler
         
         handler.PlatformView.Transform = CGAffineTransform.MakeScale(-1, 1);
         handler.PlatformView.TitleLabel.Transform = CGAffineTransform.MakeScale(-1, 1);
-        handler.PlatformView.ImageView.Transform = CGAffineTransform.MakeScale(-1, 1);
+
+        var imageViewTransform = handler.PlatformView.ImageView.Transform;
+        handler.PlatformView.ImageView.Transform = CGAffineTransform.MakeScale(-1 * imageViewTransform.A, 1 * imageViewTransform.D);
     }
     
     private static async partial void OverrideMapImageSource(ButtonHandler handler, Button button)
@@ -42,6 +52,9 @@ public partial class ButtonHandler : Microsoft.Maui.Handlers.ButtonHandler
         
         handler.PlatformView.SetImage(handler.PlatformView.CurrentImage.ImageWithRenderingMode(UIImageRenderingMode.AlwaysTemplate), UIControlState.Normal);
         handler.PlatformView.ImageView.TintColor = button.ImageTintColor.ToPlatform();
+        handler.PlatformView.ImageView.ContentMode = UIViewContentMode.ScaleAspectFit;
+
+        
     }
 
     private static partial void MapAdditionalHitBoxSize(ButtonHandler handler, Button button)