+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [17.11.0] 
+- Changed touch effect color on Android for dui:Buttons
+- Icon size on buttons will now scale down with the Button's height
+
 ## [17.8.0] 
 - Added Button styles